Extract preflight handler to a class
